refactor: move config to internal package
