+++ cmd/auth-service/container/cache.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	"github.com/DucTran999/auth-service/config"
+	"github.com/DucTran999/auth-service/internal/config"
 	"github.com/DucTran999/auth-service/pkg"
 	"github.com/DucTran999/cachekit"
 	"github.com/DucTran999/shared-pkg/logger"
