+++ internal/server/http_server.go
@@ -1,7 +1,7 @@
 package server
 
 import (
-	"github.com/DucTran999/auth-service/config"
+	"github.com/DucTran999/auth-service/internal/config"
 	"github.com/DucTran999/auth-service/internal/gen"
 	"github.com/DucTran999/shared-pkg/server"
 )
