+++ internal/worker/session_cleaner.go
@@ -6,7 +6,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/DucTran999/auth-service/config"
+	"github.com/DucTran999/auth-service/internal/config"
 	"github.com/DucTran999/auth-service/internal/handler/background"
 	"github.com/DucTran999/shared-pkg/logger"
 )
